+++ src/style-spec/expression/expression.test.ts
@@ -1,6 +1,6 @@
 import {createPropertyExpression, Feature, GlobalProperties, StylePropertyExpression} from '../expression';
 import definitions from './definitions';
-import v8 from '../reference/v8.json';
+import v8 from '../reference/v8.json' assert {type: 'json'};
 import {StylePropertySpecification} from '../style-spec';
 import ParsingError from './parsing_error';