+++ src/style-spec/validate/validate_terrain.test.ts
@@ -1,5 +1,5 @@
 import validateTerrain from './validate_terrain';
-import v8 from '../reference/v8.json';
+import v8 from '../reference/v8.json' assert {type: 'json'};
 
 describe('Validate Terrain', () => {
     test('Should return error in case terrain is not an object', () => {